+++ modules/network_synchronizer/godot_backward_utility_cpp.h
@@ -5,7 +5,7 @@
 	@author AndreaCatania
 */
 
-#include "core/engine.h"
+#include "core/config/engine.h"
 #include "scene/main/viewport.h"
 #define FLOAT REAL
 #define STRING_NAME STRING