+++ modules/network_synchronizer/net_action.cpp
@@ -1,6 +1,6 @@
 #include "net_action.h"
 
-#include "core/error_macros.h"
+#include "core/error/error_macros.h"
 #include "core/os/os.h"
 #include "scene/main/node.h"
 #include "scene_synchronizer.h"