+++ modules/network_synchronizer/scene_diff.h
@@ -1,3 +1,6 @@
+#ifndef SCENE_DIFF_H
+#define SCENE_DIFF_H
+
 /*************************************************************************/
 /*  scene_diff.h                                                         */
 /*************************************************************************/
@@ -32,10 +35,7 @@
 	@author AndreaCatania
 */
 
-#ifndef SCENE_DIFF_H
-#define SCENE_DIFF_H
-
-#include "core/class_db.h"
+#include "core/object/class_db.h"
 #include "net_utilities.h"
 
 class SceneSynchronizer;