+++ platform/uwp/os_uwp.h
@@ -36,7 +36,7 @@
 #include "core/math/transform_2d.h"
 #include "core/os/input.h"
 #include "core/os/os.h"
-#include "core/ustring.h"
+#include "core/string/ustring.h"
 #include "drivers/xaudio2/audio_driver_xaudio2.h"
 #include "joypad_uwp.h"
 #include "main/input_default.h"